requirements.txt
onnxruntime-gpu
ultralytics
opencv-python
mss
pygame
pyserial
numpy